refactor: implement index.js game logic to game.js
It didn't make a lot of sense to have a Game constructor and the game logic to be housed inside index.js so I've moved the games logic and ui components (for now) to game.js.

feat: add seaBoard variables for each players sea to be drawn to, which will be used in multiple places

fix: removed playerInstance expressions as they were removed in favor of a new Player class object now being stored in players array

feat: add logic to automatically position ships for players, later i'll write a function that allows the players to position ships themselves

feat: add firstMoveModal which will be used after new game has started to tell the players to give the computer to player 1, this can be used after automatic placement of ships and when 2 human players are playing, or after the future placeShips logic is added.

feat: add buildSeaBoard which will create a visual element representing a players sea

feat: add buildAttackScreen but its a modal, and this modal is what players will see when it's their turn to attack the other player. They'll see a version of the other players sea board and they'll see their own sea board. In the future I'll create listeners on the spans so that when a user clicks on a span they will attack that sea spot (span). I might create a confirmation box as well. I'll also build a key to tell the user what the background colors and values in the spans mean. I also would like to add a past attack result box which tells the current user what the other user accomplished the previous turn.