feat: add verifyAttack modal function
This function allows the users to confirm their attack, by also using a style on the sea spot to highlight it, this will alleviate any mis clicks. Included logic to disallow sea spot listeners from invoking when clicked while this modal is active, added the ability to the switch player modal as well. This function also records if confirmed the attack is a hit or miss and saves it to a lastMove variable to be used to show the other player what his opponents move resulted in. This function also continues the game loop depending on if the second player is a computer or is human, and increments a turnCount variable to tell users what turn number it is.

feat: add event listeners on sea spots

These listeners use a conditional to check the current value of the sea spot, if it is empty " " then that spot triggers a verifyAttack modal before registering their attack.

feat: add result text to switch player modal

fix: change selector for acquiring player 2 type value

feat: add isModalActive variable to disallow sea spot listeners to invoke while a modal is active

fix: change switchPlayers logic to actually work

fix: removed text color from sea spot text

feat: add computerAttack function

This function is still under development but will allow the computer to intelligently attack spots it thinks there could be ships at, and uses a variable to remember what the biggest ship is, I might rework that theory to include an arrary of ship sizes.