feat: change listener function to be async
The listener function in reference is when a user is facing the computer and the users turn is over and the computer needs to attack. The listener is changed to asynchronous as the computerAttack function takes time to operate.

feat: change computerAttack to random only attacks

I was confused that in the extra credit there was a more intelligent computer approach that we could go with, however it turned out it was just have the computer attack adjacent spots around a previous hit attack. For now I've shelved the adjacent attack, and have added a random attack which is required in the basic version of this project.

fix: change computer-attacking-modal to be more of a dynamic modal

This modal now has a result readout and a button, so the user can see what the computers attack resulted in and then can press the button to move on to attack the computer. This modal says the computer is attacking and once it is done changes the attack to lastMove which is the result of the computer attack. Human players are required to press a button to advance to their attacking screen.

fix: change turnCount to start with a value of one

Pretty self explanatory, when a users sees the count for the first time it should be player one and during their first turn, therefore this is turn 1

refactor: copy lastMove assigning code to computerAttacks

I copied over and changed values so that the computers move is saved as a lastMove so now the computers last move will be shown to the side of the human players defensive board