style: change background color for prospectives
The previous color was lackluster and pink really makes it stick out now

style: change position of verify attack modal

Changed the positioning from absolute to fixed, fixed allows the modal to move with the viewport scroll.

style: add rule for computer attack modal

This modal uses the same styling as verify-attack-modal however with a slightly different background and a zindex so hopefully if it ever appears wtih the verify attack modal, which it shouldnt, it will overlap it.