feat: add gameEndCheck function
gameEndCheck will be called after a human or computer has made an attack and will evaluate if the game is over depending on how many boats the defending player of each attack has boats that are not sunken. For now this function just console.logs a statement, later it will render an end game modal and remove all game listener elements on the 'board'.

refactor: change gameState to gameInstance

feat: add gameState variable

This variable will now keep track of if a game exists or is completed, and will invoke the function to create the game winner board.

fix: change buildSeaBoard conditionals

Changed the argument to the conditions from checking for name equality which could cause game boards to render incorrectly if the players were the same name, to now using player object equality.