feat: fix gameEndCheck function
This function will be used to check if either player has any ships left and if true triggers a game over element to be made and displayed to the users.

feat: add gameOver function

This function will be used by gameEndCheck, and is the function that creates the game over element which features how many turns the game took to complete, what each players sea board looked like at game completion, a key reminding users what each sea spots current background color and text content means, and a button that will create a new game and wipe the old game from the dom.

feat: add buildFinalSeaBoard function

This function is used by gameOver to create elements that represent the game boards of each player, it takes a player object as argument and builds a board no matter if that player was the attacker or defender (which the initial sea board function did take in to affect how the board would be built).

fix: change dynamic player name for switchPlayer

This function was actually showing the wrong name, it was showing the previous attackers name and now shows the next attackers name.

feat: add attackers name to lastMove

Previously lastMove just said "Attack was a miss at y,x", now it's prefaced with the attacking players name.