style: change switch player modal styles
Enlarged this modal and gave it a background color to pop out more, as the previous white was mundane. Also enlarged the height and width so players can't simply ignore it.

style: adjust headers to boards

These headers weren't centered and felt as if they could have used some margins.

style: add rules for winSummary and children

winSummary creates the game over information screen and contains many child elements, one which needed a grid and it's children to be positioned within that grid.

style: add hover pseudo selector for sea spots

In addition to the previously added clicked pseudo selector which changes the background color for a clicked sea spot, I've added a hover so that as the user navigates over sea spots the targeted spot by the cursor will have their background color changed. I used the same color as the clicked pseudo selector as I thought it made sense and looked decent enough.