feat: add replaceShip
This method will allow users when placing their ships to reposition ships.

feat: added type to previously existing methods

type allows us to target specific ships, and is needed for placement and replacement purposes.

feat: add a return statement when a ship cannot be placed

This statement will be shown to the user via pseudo selector below the ship they are trying to place or replace/etc.