refactor: change _boats array initial value
_boats initial value was an empty array, we needed to change it to have 5 elements so that we can dynamically alter specific elements at any time, this change gives us that ability.

refactor: change sea references

Before this change we referenced "sea" not "_sea", which is wrong because we need to keep in mind that sea is a private property, and sea would be a getter return and not technically used to set values.

fix: change expressions that target dynamic index's of _sea

There are some expressions that when evaluated should access a dynamic index, yet those expressions were incorrect. These can be summarized as when looking for a prospective block of sea spots to place a future ship we need to correctly target the prospective sea spots but the expressions before this change were targeting the wrong prospective sea spots by one square. There were also some instances of targeting the wrong direction, for example when placing a ship in an 'up' or 'down' orientation that  operator used was incorrect in both cases causing the ship to be placed the other orientation. Likewise operators in _vacancyChecker were changed as they were incorrect as well.

refactor: add else condition in placeShips

In the future I'll tell the user why an action of theirs wasn't possible, so this is an initial attempt at creating that message.

refactor: change how boats are added to _boats

This method now uses a switch statement to create Ship's inside the _boats array at specific index's, as mentioned in another change to _boats. This helps set specific types of boats at specific index's so that other methods can access the correct ship. Similarly I've changed how values are added to _sea, which is now related to the index of the ship in _boats which ties together the type of the ship and the value representing it in the seaboards.

refactor: change boatsRemaining to use equality check

With the change to _boats initial values, we now check for a specific value to the sunk property, before it was a check to see if sunk was truthy which does not work now that _boat elements can be empty strings.

refactor: change parameter names for replaceShip

This change was needed because we need to differentiate between the old ship and the new ship.

refactor: change replaceShip logic

This method has been rewritten as it didn't work after changes to _boats, etc. The statement that leads to removing the old boat was not working correctly and that's because I forgot sea is an array of arrays and if i'm going to use map to change it then i need to chain a map within a map so that it runs for each row of arrays in the sea array.

refactor: change clearShips method

This method earlier changed the value the _boats property of an instance of this class pointed to, to an empty array. This now points to an array of 5 empty strings, which is useful to us because of the other changes to this file and methods that work on index's related to types of a ships.