style: add rule for place-ships-ui
In addition to the ui for placing ships, added rules for smaller components

style: add rule for verifyModal

This element is used project wide not just in the place ships ui, it needs to overlap ui's so a fixed position is needed.

style: add rule for player name spans

I'd like to work this element in around any mention of a players name, mainly on the switch player modal.

Had added invalid pseudo class rules for invalid input but decided to rely on error messages being shown to users to be good enough for now.