fix: change withSeaRange to be dynamic
Earlier version of withinSeaRange used simple conditionals to check if the first sea spot and last sea spot of a prospective new ship were within the sea boards range, but to make placeShip not so dense I've gone ahead and moved the logic of calculating the required range for a new ship into withnSeaRange.

fix: change placeShip to not rely on conditions

Earlier version of placeShip used 4 different conditions to  call withinSeaRange, but withinSeaRange was changed to be more advanced and bakes in the orientations inside it's function.

fix: change vacancyChecker

This methods previous version didn't actually work correctly, it attempted to make temp arrays and then check the values of those elements now this method skips making a temp array and correctly checks the prospective sea spots values of where the new ship will be placed.

fix: move counter incrementer inside _addShip

The previous version was incorrect and was incrementing the counter continually during each iteration, not after each iteration.

fix: add missing arguments to withinSeaRange inside replaceShip method

The past version was missing orientation and length and thus was allowing ships that could not fit inside the sea board. Also fixed an expression which was incorrectly placing a boat in the wrong orientation