feat: add condition to placeShip
Added an else condition that returns false, as this will now invoke inputVacancyErrorDisplay in the src/game.js file when an unplaced ship cannot be placed because of another ship occupying the desired sea spots of this new ship.

fix: revert changes to vacancyChecker

I confused targeted value of a sea spot with the boat number of a prospective boat to be placed, and the result was a rhetorical condition, which would always be true.