feat: add placeShips function
This function dynamically invokes the ui for players to place ships depending on how many human players there are, and thus has its own series of dynamic invocations that are awaited.

feat: add placeShipsModal

This function creates a ui that allows players to place their ships into their sea board, using either a random button or a table with inputs, users can be specific or not if they choose.

feat: add verifyModal

This function which I made in a previous side project, creates a pop up confirmation box which will use project wide to confirm game changing decisions by players. This function is awaited when invoked meaning it needs to return a value or a promise, and we have a confirm button that if clicked resolves that promise and likewise a cancel button that resolves the promise

fix: adjusted game logic loop

The core game logic of players attacking each other, or a computer, and eventually being alerted that the game is over is stored within the confirm attack listener. I'm not sure if that's the best place for it but it is where I've built it.

fix: remove not operator prefacing expressions

The not operator I learned again, short circuits anything that follows the expression that follows it, therefore whatever you have written after the expression is ignored basically, so here I've changed !isModalActive to isModalActive !== true.

fix: rewrite switchPlayer function to return a promise

switchPlayer is awaited and because it creates a ui which needs to return a value or a promise; I learned that event listener functions do not return, clicking the switch button resolves(true) which meets await'ed requirement.

feat: add to buildSeaBoard dynamic classes

buildSeaBoard is used in multiple places and for the ability to show users where their ships are being placed I've added classes for specific ships, and in the css these different ships have background colors as well as the row for input fields in the placeShips ui so users can visually relate what input fields are for what ship they just placed.

feat: add computerTargetedAttack

Previously we used computerRandomlyAttacks, but the extra credit mentioned that we could build a function to attack spots adjacent to successfully attacked sea spots, so this method does just that. If a successful hit does not exist yet then the prior attack method of random is used, otherwise if a hit is found this functions checks in order what adjacent spot has not been attacked starting from the spot on its left, below it, before it, above it, and if any are not attacked then invokes the receiveAttack function on that player, otherwise continues to iterate over sea spots looking for hits. If a hit is not found after the loop has finished, it invokes the random attack function. This function and the random attack function return true which the await prefacing these functions receives and resolves.

refactor: add randomlyPlaceShips function

This function existed before but was baked into the program, now has been refactored into its own function. It has been rewritten as the previous version had an unnecessary parent while loop that would iterate until the player had 5 ships.

feat: add inputVacancyErrorDisplay to addOrReplaceShip

This function could not be invoked in the higher up the chain condition that is checked when any of the add/replace boat buttons are clicked on because of the way vacancyChecker operates on the current sea of a players gameboard. I would have had to change many gameboard.js methods or copied over most of vacancyChecker's code into a new function in game.js, so instead I found the two scopes where I could just call this function to output an error.